Some better results. after_start_complete test works, slash produces a slash and spends it, but we detect it as no-reward and should have a reward.
