+++ clsp/onchain/referee.clsp
@@ -8,6 +8,7 @@
 (import std.condition_codes)
 (import std.match)
 (import std.li)
+(import std.print)
 (import onchain.game_codes)
 
 ;
@@ -78,19 +79,19 @@
                     VALIDATION_INFO_HASH)))
             conditions (a mover_puzzle solution)
             (assert
-                VALIDATION_INFO_HASH
-                (<= (strlen new_move) MAX_MOVE_SIZE)
-                (<= new_mover_share AMOUNT)
-                (>= new_mover_share 0)
-                (= 32 (strlen new_validation_info_hash))
-                (= MOVER_PUZZLE_HASH (shatree mover_puzzle))
+                (print "validation info case" VALIDATION_INFO_HASH)
+                (print "strlen case" (<= (strlen new_move) MAX_MOVE_SIZE))
+                (print "new_mover_share case" (<= new_mover_share AMOUNT))
+                (print "new_mover_share_gt case" (>= new_mover_share 0))
+                (print "new_validation_info" (= 32 (strlen new_validation_info_hash)))
+                (print "mover_puzzle_hash case" (= MOVER_PUZZLE_HASH (shatree mover_puzzle)))
                 ; Check that the child output is made
-                (match
+                (print "verify conditions case" (match
                     (lambda ((& new_puzzle_hash AMOUNT) (condname arg1 arg2))
                         (logand (= condname CREATE_COIN) (= arg1 new_puzzle_hash) (= arg2 AMOUNT))
                     )
                     conditions
-                )
+                ))
                 (li (list REMARK new_move new_validation_info_hash new_mover_share new_max_move_size) 
                         (list ASSERT_BEFORE_HEIGHT_RELATIVE TIMEOUT) &rest conditions)
             )
