+++ src/referee/their_turn.rs
@@ -400,249 +400,6 @@ impl TheirTurnReferee {
         curry_referee_puzzle_hash(allocator, &self.fixed.referee_coin_puzzle_hash, &args)
     }
 
-    // Ensure this returns
-    fn get_transaction(
-        &self,
-        allocator: &mut AllocEncoder,
-        coin_string: &CoinString,
-        always_produce_transaction: bool,
-        puzzle: Puzzle,
-        targs: &RefereePuzzleArgs<StateUpdateProgram>,
-        args: &OnChainRefereeSolution,
-    ) -> Result<Option<RefereeOnChainTransaction>, Error> {
-        let our_move = self.is_my_turn();
-
-        let my_mover_share = if our_move {
-            targs.game_move.basic.mover_share.clone()
-        } else {
-            self.fixed.amount.clone() - targs.game_move.basic.mover_share.clone()
-        };
-
-        if always_produce_transaction || my_mover_share != Amount::default() {
-            let signature = args.get_signature().unwrap_or_default();
-
-            // The transaction solution is not the same as the solution for the
-            // inner puzzle as we take additional move or slash data.
-            //
-            // OnChainRefereeSolution encodes this properly.
-            let transaction_solution = args.to_clvm(allocator).into_gen()?;
-            debug!("transaction solution inputs {args:?}");
-            let transaction_bundle = Spend {
-                puzzle: puzzle.clone(),
-                solution: Program::from_nodeptr(allocator, transaction_solution)?.into(),
-                signature,
-            };
-            let output_coin_string = CoinString::from_parts(
-                &coin_string.to_coin_id(),
-                &puzzle.sha256tree(allocator),
-                &my_mover_share,
-            );
-            return Ok(Some(RefereeOnChainTransaction {
-                bundle: transaction_bundle,
-                amount: self.fixed.amount.clone(),
-                coin: output_coin_string,
-            }));
-        }
-
-        // Zero mover share case.
-        Ok(None)
-    }
-
-    /// Output coin_string:
-    /// Parent is hash of current_coin
-    /// Puzzle hash is my_referee_puzzle_hash.
-    ///
-    /// Timeout unlike other actions applies to the current ph, not the one at the
-    /// start of a turn proper.
-    pub fn get_transaction_for_timeout(
-        &self,
-        allocator: &mut AllocEncoder,
-        coin_string: &CoinString,
-    ) -> Result<Option<RefereeOnChainTransaction>, Error> {
-        debug!("get_transaction_for_timeout turn {}", self.is_my_turn());
-        debug!(
-            "mover share at start of action   {:?}",
-            self.args_for_this_coin().game_move.basic.mover_share
-        );
-        debug!(
-            "mover share at end   of action   {:?}",
-            self.spend_this_coin().game_move.basic.mover_share
-        );
-
-        let targs = self.spend_this_coin().neutralize();
-        let puzzle = curry_referee_puzzle(allocator, &self.fixed.referee_coin_puzzle, &targs)?;
-
-        self.get_transaction(
-            allocator,
-            coin_string,
-            false,
-            puzzle,
-            &targs,
-            &OnChainRefereeSolution::Timeout,
-        )
-    }
-
-    /// The move transaction works like this:
-    ///
-    /// The referee puzzle has the hash of the puzzle of another locking coin,
-    /// possibly the standard coin, and uses that to secure against another person
-    /// commanding it.  This isn't the be confused with the coin that serves as the
-    /// parent of the referee coin which is also assumed to be a standard puzzle
-    /// coin.
-    ///
-    /// The inner coin, assuming it is a standard coin, takes the puzzle reveal
-    /// for the above puzzle and the solution for that inner puzzle as the last two
-    /// arguments to the move case of how it's invoked.
-    ///
-    /// The output conditions to step it are therefore built into those conditions
-    /// which needs to include the puzzle hash of the target state of the referee
-    /// (their move, the state precipitated by our move set as the current game
-    /// state).
-    ///
-    /// We do the spend of the inner puzzle to that puzzle hash to progress the
-    /// referee coin.
-    ///
-    /// One consequence of this is that we must sign it with the synthetic private
-    /// key as the standard puzzle embeds a synthetic public key based on it.
-    ///
-    /// In all cases, we're spending a referee coin that already exists.  The use
-    /// of the mover coin here is purely to take advantage of its puzzle to provide
-    /// a signature requirement.
-    pub fn get_transaction_for_move(
-        &self,
-        allocator: &mut AllocEncoder,
-        coin_string: &CoinString,
-        on_chain: bool,
-    ) -> Result<RefereeOnChainTransaction, Error> {
-        // We can only do a move to replicate our turn.
-        assert!(self.processing_my_turn());
-        let target_args = self.spend_this_coin();
-        let spend_puzzle = self.on_chain_referee_puzzle(allocator)?;
-
-        // Get the puzzle hash for the next referee state.
-        // This reflects a "their turn" state with the updated state from the
-        // game handler returned by consuming our move.  This is assumed to
-        // have been done by consuming the move in a different method call.
-
-        // Get the current state of the referee on chain.  This reflects the
-        // current state at the time the move was made.
-        // The current referee uses the previous state since we have already
-        // taken the move.
-        //
-        debug!("get_transaction_for_move: previous curry");
-        let args = self.args_for_this_coin();
-
-        //debug!("transaction for move: state {:?}", self.state);
-        debug!("get_transaction_for_move: source curry {args:?}");
-        debug!("get_transaction_for_move: target curry {target_args:?}");
-        assert_eq!(
-            target_args.mover_puzzle_hash,
-            self.fixed.my_identity.puzzle_hash
-        );
-        // assert_ne!(args.mover_puzzle_hash, target_args.mover_puzzle_hash);
-        // assert_eq!(args.mover_puzzle_hash, target_args.waiter_puzzle_hash);
-        assert!(matches!(
-            self.state.borrow(),
-            TheirTurnRefereeMakerGameState::AfterOurTurn { .. }
-        ));
-        assert!(matches!(
-            self.get_game_handler(),
-            GameHandler::TheirTurnHandler(_)
-        ));
-
-        if let Some((_, ph, _)) = coin_string.to_parts() {
-            if on_chain {
-                let start_ph = curry_referee_puzzle_hash(
-                    allocator,
-                    &self.fixed.referee_coin_puzzle_hash,
-                    &args,
-                )?;
-                let end_ph = curry_referee_puzzle_hash(
-                    allocator,
-                    &self.fixed.referee_coin_puzzle_hash,
-                    &target_args,
-                )?;
-                debug!("spend puzzle hash {ph:?}");
-                debug!("this coin start {start_ph:?}");
-                debug!("this coin end   {end_ph:?}");
-                // assert_eq!(ph, start_ph);
-            }
-        }
-
-        assert_eq!(
-            Some(&args.game_move.validation_info_hash),
-            target_args.previous_validation_info_hash.as_ref()
-        );
-        // debug!(
-        //     "transaction for move: from {:?} to {target_args:?}",
-        //     self.args_for_this_coin()
-        // );
-        let target_referee_puzzle_hash = curry_referee_puzzle_hash(
-            allocator,
-            &self.fixed.referee_coin_puzzle_hash,
-            &target_args,
-        )?;
-        let target_referee_puzzle =
-            curry_referee_puzzle(allocator, &self.fixed.referee_coin_puzzle, &target_args)?;
-        assert_eq!(
-            target_referee_puzzle.sha256tree(allocator),
-            target_referee_puzzle_hash
-        );
-
-        let inner_conditions = [(
-            CREATE_COIN,
-            (
-                target_referee_puzzle_hash.clone(),
-                (self.fixed.amount.clone(), ()),
-            ),
-        )]
-        .to_clvm(allocator)
-        .into_gen()?;
-
-        // Generalize this once the test is working.  Move out the assumption that
-        // referee private key is my_identity.synthetic_private_key.
-        debug!("referee spend with parent coin {coin_string:?}");
-        debug!(
-            "signing coin with synthetic public key {:?} for public key {:?}",
-            self.fixed.my_identity.synthetic_public_key, self.fixed.my_identity.public_key
-        );
-        let referee_spend = standard_solution_partial(
-            allocator,
-            &self.fixed.my_identity.synthetic_private_key,
-            &coin_string.to_coin_id(),
-            inner_conditions,
-            &self.fixed.my_identity.synthetic_public_key,
-            &self.fixed.agg_sig_me_additional_data,
-            false,
-        )?;
-
-        let args_list = OnChainRefereeSolution::Move(OnChainRefereeMove {
-            details: target_args.game_move.clone(),
-            max_move_size: target_args.max_move_size,
-            mover_coin: IdentityCoinAndSolution {
-                mover_coin_puzzle: self.fixed.my_identity.puzzle.clone(),
-                mover_coin_spend_solution: referee_spend.solution.p(),
-                mover_coin_spend_signature: referee_spend.signature.clone(),
-            },
-        });
-
-        if let Some(transaction) = self.get_transaction(
-            allocator,
-            coin_string,
-            true,
-            spend_puzzle,
-            &target_args.neutralize(),
-            &args_list,
-        )? {
-            Ok(transaction)
-        } else {
-            // Return err
-            Err(Error::StrErr(
-                "no transaction returned when doing on chain move".to_string(),
-            ))
-        }
-    }
-
     /// Run the initial validator for a their turn move.  We must run the their turn validator
     /// first before we run the turn handler to provide a new state to the turn handler.
     pub fn run_state_update(
@@ -945,7 +702,6 @@ impl TheirTurnReferee {
         match &result.original {
             TheirTurnResult::Slash(evidence) => {
                 // Slash specified.
-                todo!();
                 let args = self.spend_this_coin();
                 let slash_spend = self.make_slash_spend(allocator, coin_string)?;
                 let new_puzzle =
@@ -1047,7 +803,6 @@ impl TheirTurnReferee {
         // after the transaction below has been spent so its
         // parent is the coin id of that coin.
         let current_mover_share = self.get_our_current_share();
-        todo!();
 
         let (state, validation_program) = self.get_validation_program_for_their_move()?;
         let reward_amount = self.fixed.amount.clone() - current_mover_share;
